refactor: SVG 폴더 정리
