+++ src/assets/svgs/bottomNav/index.ts
@@ -1,0 +1,2 @@
+export { default as IcDeliver } from './IcDeliver';
+export { default as IcLocation } from './IcLocation';
