+++ src/assets/svgs/header/ImgHeaderLogo.tsx
