+++ src/assets/svgs/header/index.ts
@@ -1,0 +1,3 @@
+export { default as BtnMenu } from './BtnMenu';
+export { default as CmpBtnBack } from './CmpBtnBack';
+export { default as ImgHeaderLogo } from './ImgHeaderLogo';
