+++ src/assets/svgs/index.ts
@@ -8,11 +8,3 @@ export { default as ImgMenu6 } from './burger/ImgMenu6';
 export { default as ImgMenu7 } from './burger/ImgMenu7';
 export { default as ImgMenu8 } from './burger/ImgMenu8';
 export { default as ImgMenu9 } from './burger/ImgMenu9';
-export { default as BtnAccordion } from './BtnAccordion';
-export { default as BtnMenu } from './BtnMenu';
-export { default as BtnX } from './BtnX';
-export { default as CmpBtnBack } from './CmpBtnBack';
-export { default as IcSearch } from './IcSearch';
-export { default as ImgHeaderLogo } from './ImgHeaderLogo';
-export { default as IcDeliver } from './bottomNav/IcDeliver';
-export { default as IcLocation } from './bottomNav/IcLocation';
