+++ src/assets/svgs/sidebar/BtnAccordion.tsx
