+++ src/assets/svgs/sidebar/IcSearch.tsx
