+++ src/assets/svgs/sidebar/index.ts
@@ -1,0 +1,3 @@
+export { default as BtnAccordion } from './BtnAccordion';
+export { default as BtnX } from './BtnX';
+export { default as IcSearch } from './IcSearch';
